# swagger-codegen-filepart-sample/src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: File Upload API Sample
  description: Sample API demonstrating reactive file upload with FilePart
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server

paths:
  /files:
    post:
      summary: Upload a file
      description: Upload a single file with metadata using reactive FilePart
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
            encoding:
              metadata:
                contentType: application/json
              file:
                contentType: application/octet-stream
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /files/{fileId}:
    get:
      summary: Get file metadata
      description: Retrieve metadata for a specific file
      operationId: getFileMetadata
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File metadata retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadata'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    FileUploadRequest:
      type: object
      required:
        - file
        - fileName
        - uploadedBy
      properties:
        file:
          type: string
          format: binary
          description: The file to upload
        fileName:
          type: string
          description: Name of the file
          minLength: 1
          maxLength: 255
        uploadedBy:
          type: string
          description: User ID who uploaded the file
        metadata:
          $ref: '#/components/schemas/FileMetadata'
        tags:
          type: array
          items:
            type: string
          description: Optional tags for the file

    FileMetadata:
      type: object
      properties:
        fileId:
          type: string
          format: uuid
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        contentType:
          type: string
        uploadedBy:
          type: string
        uploadedAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string

    FileUploadResponse:
      type: object
      properties:
        fileId:
          type: string
          format: uuid
        message:
          type: string
        uploadedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time